ta - clean-up of folder structure
