--- app/actions.js
@@ -1,6 +1,0 @@
-var flux = require('flux-react');
-
-module.exports = flux.createActions([
-  'upsertTrack',
-  'deleteTrack'
-]);
